requirements.txt
django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2==2.8.4
google-cloud-logging==1.15.0